Issue #4575: Modified ClassFanOutComplexityCheckTest.java and moved its input files to the classfanoutcomplexity subdirectory
